+++ lib/fog/radosgw/usage.rb
@@ -1,4 +1,3 @@
-require 'fog/radosgw/core'
 require 'time'
 
 module Fog
